refactor(qwen3_5): move Qwen3.5 VL wrapper into qwen_vl.py
RBLNOptimumQwen3_5ForConditionalGeneration is a small subclass of the Qwen2.5-VL
wrapper, so keep it alongside the other Qwen-VL wrappers instead of in its own
file. No behavior change; the auto-discovery import just moves to the qwen_vl
block.
